LayoutTests/imported/w3c/web-platform-tests/css/css-counter-styles/idlharness-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface CSSRule: original interface defined
PASS Partial interface CSSRule: member names are unique
FAIL CSSCounterStyleRule interface: existence and properties of interface object assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface object length assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface object name assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute name assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute system assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute symbols assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute additiveSymbols assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute negative assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute prefix assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute suffix assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute range assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute pad assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute speakAs assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule interface: attribute fallback assert_own_property: self does not have own property "CSSCounterStyleRule" expected property "CSSCounterStyleRule" missing
FAIL CSSCounterStyleRule must be primary interface of counter assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of counter assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "name" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "system" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "symbols" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "additiveSymbols" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "negative" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "prefix" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "suffix" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "range" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "pad" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "speakAs" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSCounterStyleRule interface: counter must inherit property "fallback" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSRule interface: counter must inherit property "COUNTER_STYLE_RULE" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CSSRule interface: constant COUNTER_STYLE_RULE on interface object assert_own_property: expected property "COUNTER_STYLE_RULE" missing
FAIL CSSRule interface: constant COUNTER_STYLE_RULE on interface prototype object assert_own_property: expected property "COUNTER_STYLE_RULE" missing
